last-minute fixes
- shuffle some pins arround to fix temp sensor
- pmu_get_battery_voltage() now outputs float w/ correct units
- sd no longer allows custom filenames; generates them as needed
- fix modem init disappearing for some reason
- add millis() timestamp to data log
- support seperate motor1/motor2 enable pins
- fix temp sensor blocking madness

+++ include/pmu.h
@@ -12,4 +12,4 @@ bool pmu_init();
  */
 
 int pmu_get_battery_percent();
-int pmu_get_battery_voltage();
+float pmu_get_battery_voltage();
